v2.0.0 サブコマンド名を改名し一貫性を改善
- claude-skill → up claude-skill（upのネストサブコマンドに移動）
- load → scan（既存構造のスキャン・取り込みを明確に表現）
- upコマンドにclaude-skillの生成処理を統合し、claude_skill.rsを削除
- README.mdのバージョン情報を更新
