+++ CLAUDE.md
@@ -32,7 +32,7 @@ cargo run -- up           # コード生成
 
 4層構造:
 
-- **`src/cli/`** - CLIコマンド層（clap）。`new`, `up`, `claude-skill`, `completion`サブコマンド
+- **`src/cli/`** - CLIコマンド層（clap）。`new`, `up`（`up claude-skill`含む）, `scan`, `rm`, `completion`サブコマンド
 - **`src/project_management/config/`** - YAML設定の解析（`parser.rs`）、バリデーション（`validator.rs`）、データモデル（`models.rs`）
 - **`src/code_generation/core/`** - コード生成エンジン。`generator.rs`がエントリーポイントで、`directory_builder.rs`と`file_builder.rs`に委譲
 - **`src/code_generation/language/`** - 言語別の生成処理（`rust/`, `go/`, `python/`, `typescript/`, `javascript/`, `any/`）
@@ -62,3 +62,4 @@ moliの生成エンジンは3層のファイル保護を実装しており、コ
 ```
 
 複数プロジェクトを1つのmoli.ymlで管理可能。Rustの場合はワークスペース構成を自動生成。
+
