+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "moli"
-version = "1.2.1"
+version = "2.0.0"
 edition = "2021"
 
 [[bin]]
