+++ src/cli/command/mod.rs
@@ -2,8 +2,7 @@
 pub mod new;
 pub mod up;
 pub mod rm;
-pub mod load;
-pub mod claude_skill;
+pub mod scan;
 pub mod completion;
 // end auto exported by moli.
